Fixed a possible crash when logging mail
Former-commit-id: 52462c537ed475aae71c897e962b4857b1e1bd06

+++ src/game/Corpse.cpp
@@ -181,7 +181,7 @@ bool Corpse::LoadFromDB(uint32 guid, Field *fields)
     }
 
     m_time             = time_t(fields[6].GetUInt64());
-    m_type             = CorpseType(fields[7].GetUInt32());
+    m_type             = CorpseType(fields[7].GetUInt8());
     if(m_type >= MAX_CORPSE_TYPE)
     {
         TC_LOG_ERROR("FIXME","ERROR: Corpse (guidlow %d, owner %d) have wrong corpse type, not load.",GetGUIDLow(),GUID_LOPART(GetOwnerGUID()));
